+++ code/main_omp_sparse/SRKABWOR_csr_error.C
@@ -114,7 +114,7 @@ int main (int argc, char *argv[]) {
 	long long avg_it = 0;
 	bool solution_found;
 
-	int block_begin;
+	long long block_begin;
 
 	for(int run = 0; run < n_runs; run++) {
 		for (int i = 0; i < N; i++) {
